Simplify barchart script, remove tasklet line plot
Add hashes, fix legend ordering

Refactor barchart script, it was more complicated then it needed to be
and trying to do two charts in one script is unnecessary.

Remove tasklet script in favor of 'asplos21/chart_dpu_vs_host.py'
