+++ runtimeplot/barchart.csv.example
@@ -1,0 +1,8 @@
+Prepare,Copy in,Run,NR_DPUS
+0.042657,0.170938,1.169508,192
+0.056753,0.197635,0.895231,256
+0.074609,0.215836,0.765229,320
+0.082655,0.238283,0.635431,384
+0.092979,0.263554,0.514021,448
+0.108401,0.285356,0.506871,512
+0.121104,0.305812,0.404117,576
